Create Teacher Database, and moved some class methods related to students and teachers into their respective database
