+++ app.py
@@ -26,6 +26,7 @@ from routes.exam_routes.retrieve_exam_results_route import retrieve_exam_results
 from routes.exam_routes.delete_exam_result import delete_exam_result_bp
 from routes.class_routes.retrieve_student_classes_route import retrieve_student_classes_bp
 from routes.student_routes.retrieve_student_teachers_route import retrieve_student_teachers_bp
+import subprocess
 
 app = Flask(__name__)
