+++ modules/database_modules/class_database.py
@@ -94,75 +94,40 @@ class ClassesDatabase:
                 print(f"Error registering class: {error_message}") # Debugging print
                 return self.generate_response(success=False, error=error_message, status_code=500, error_code=e.pgcode)
 
-    def retrieve_teacher_classes(self, teacher_id):
-        if not teacher_id:
-            return self.generate_response(success=False, error='Teacher ID must be provided.', status_code=400)
+    def retrieve_class_students(self, class_id):
+        if not class_id:
+            return self.generate_response(success=False, error='Class ID must be provided.', status_code=400)
 
         with db_connection(self.credentials) as conn:
             try:
                 cur = conn.cursor(cursor_factory=psycopg2.extras.DictCursor)
 
-                # Check if the teacher exists
+                # Check if the class exists
                 check_query = """
-                    SELECT 1 FROM users_teachers
-                    WHERE id = %s;
+                    SELECT 1 FROM classes
+                    WHERE class_id = %s;
                 """
-                cur.execute(check_query, (teacher_id,))
-                existing_teacher = cur.fetchone()
-                if not existing_teacher:
-                    return self.generate_response(success=False, error='Teacher not found.', status_code=404)
+                cur.execute(check_query, (class_id,))
+                existing_class = cur.fetchone()
+                if not existing_class:
+                    return self.generate_response(success=False, error='Class not found.', status_code=404)
 
-                # Retrieve the classes taught by the teacher
+                # Retrieve the students registered in the class
                 query = """
-                    SELECT * FROM classes
-                    WHERE teacher_id = %s;
+                    SELECT users_students.id, name, username, email, faculty, matnum
+                    FROM users_students
+                    JOIN classes_students ON users_students.id = classes_students.student_id
+                    WHERE class_id = %s;
                 """
-                cur.execute(query, (teacher_id,))
-                classes = cur.fetchall()
+                cur.execute(query, (class_id,))
+                students = cur.fetchall()
                 cur.close()
-                classes_dict = [dict(row) for row in classes]
-                return self.generate_response(success=True, error=None, status_code=200, data=classes_dict)
+                students_dict = [dict(row) for row in students]
+                return self.generate_response(success=True, error=None, status_code=200, data=students_dict)
 
             except psycopg2.Error as e:
                 error_message = e.pgerror if e.pgerror else str(e)
-                print(f"Error retrieving classes: {error_message}")
-                return self.generate_response(success=False, error=error_message, status_code=500, error_code=e.pgcode)
-
-    def retrieve_student_classes(self, student_id):
-        if not student_id:
-            return self.generate_response(success=False, error='Student ID must be provided.', status_code=400)
-
-        with db_connection(self.credentials) as conn:
-            try:
-                cur = conn.cursor(cursor_factory=psycopg2.extras.DictCursor)
-
-                # Check if the student exists
-                check_query = """
-                    SELECT 1 FROM users_students
-                    WHERE id = %s;
-                """
-                cur.execute(check_query, (student_id,))
-                existing_student = cur.fetchone()
-                if not existing_student:
-                    return self.generate_response(success=False, error='Student not found.', status_code=404)
-
-                # Retrieve the classes attended by the student
-                query = """
-                    SELECT c.*, ut.name AS teacher_name
-                    FROM classes_students cs
-                    JOIN classes c ON cs.class_id = c.class_id
-                    JOIN users_teachers ut ON c.teacher_id = ut.id
-                    WHERE cs.student_id = %s;
-                """
-                cur.execute(query, (student_id,))
-                classes = cur.fetchall()
-                cur.close()
-                classes_dict = [dict(row) for row in classes]
-                return self.generate_response(success=True, error=None, status_code=200, data=classes_dict)
-
-            except psycopg2.Error as e:
-                error_message = e.pgerror if e.pgerror else str(e)
-                print(f"Error retrieving classes: {error_message}")
+                print(f"Error retrieving class students: {error_message}")
                 return self.generate_response(success=False, error=error_message, status_code=500, error_code=e.pgcode)
 
     def update_class(self, class_id, **kwargs):
@@ -305,42 +270,6 @@ class ClassesDatabase:
                 conn.rollback()
                 error_message = e.pgerror if e.pgerror else str(e)
                 print(f"Error adding student to class: {error_message}")
-                return self.generate_response(success=False, error=error_message, status_code=500, error_code=e.pgcode)
-
-    def retrieve_class_students(self, class_id):
-        if not class_id:
-            return self.generate_response(success=False, error='Class ID must be provided.', status_code=400)
-
-        with db_connection(self.credentials) as conn:
-            try:
-                cur = conn.cursor(cursor_factory=psycopg2.extras.DictCursor)
-
-                # Check if the class exists
-                check_query = """
-                    SELECT 1 FROM classes
-                    WHERE class_id = %s;
-                """
-                cur.execute(check_query, (class_id,))
-                existing_class = cur.fetchone()
-                if not existing_class:
-                    return self.generate_response(success=False, error='Class not found.', status_code=404)
-
-                # Retrieve the students registered in the class
-                query = """
-                    SELECT users_students.id, name, username, email, faculty, matnum
-                    FROM users_students
-                    JOIN classes_students ON users_students.id = classes_students.student_id
-                    WHERE class_id = %s;
-                """
-                cur.execute(query, (class_id,))
-                students = cur.fetchall()
-                cur.close()
-                students_dict = [dict(row) for row in students]
-                return self.generate_response(success=True, error=None, status_code=200, data=students_dict)
-
-            except psycopg2.Error as e:
-                error_message = e.pgerror if e.pgerror else str(e)
-                print(f"Error retrieving class students: {error_message}")
                 return self.generate_response(success=False, error=error_message, status_code=500, error_code=e.pgcode)
 
     def del_student_from_class(self, student_id, class_id):
