+++ modules/database_modules/student_database.py
@@ -38,7 +38,6 @@ class StudentDatabase:
 
     # TODO retrieve_student_data method
 
-        # python
     def retrieve_student_teachers(self, student_id):
         if not student_id:
             return self.generate_response(success=False, error='Student ID must be provided', status_code=400)
@@ -76,6 +75,43 @@ class StudentDatabase:
             except psycopg2.DatabaseError as e:
                 error_message = e.pgerror if e.pgerror else str(e)
                 print('Error retrieving teachers:', error_message)
+                return self.generate_response(success=False, error=error_message, status_code=500, error_code=e.pgcode)
+
+    def retrieve_student_classes(self, student_id):
+        if not student_id:
+            return self.generate_response(success=False, error='Student ID must be provided.', status_code=400)
+
+        with db_connection(self.credentials) as conn:
+            try:
+                cur = conn.cursor(cursor_factory=psycopg2.extras.DictCursor)
+
+                # Check if the student exists
+                check_query = """
+                    SELECT 1 FROM users_students
+                    WHERE id = %s;
+                """
+                cur.execute(check_query, (student_id,))
+                existing_student = cur.fetchone()
+                if not existing_student:
+                    return self.generate_response(success=False, error='Student not found.', status_code=404)
+
+                # Retrieve the classes attended by the student
+                query = """
+                    SELECT c.*, ut.name AS teacher_name
+                    FROM classes_students cs
+                    JOIN classes c ON cs.class_id = c.class_id
+                    JOIN users_teachers ut ON c.teacher_id = ut.id
+                    WHERE cs.student_id = %s;
+                """
+                cur.execute(query, (student_id,))
+                classes = cur.fetchall()
+                cur.close()
+                classes_dict = [dict(row) for row in classes]
+                return self.generate_response(success=True, error=None, status_code=200, data=classes_dict)
+
+            except psycopg2.Error as e:
+                error_message = e.pgerror if e.pgerror else str(e)
+                print(f"Error retrieving classes: {error_message}")
                 return self.generate_response(success=False, error=error_message, status_code=500, error_code=e.pgcode)
 
     # Private method to generate a consistent JSON response
