+++ routes/class_routes/retrieve_student_classes_route.py
@@ -1,9 +1,9 @@
 from flask import Blueprint, request, jsonify
-from modules.database_modules.class_database import ClassesDatabase
+from modules.database_modules.student_database import StudentDatabase
 
 
 retrieve_student_classes_bp = Blueprint('retrieve_student_classes', __name__)
-db = ClassesDatabase()
+db = StudentDatabase()
 
 
 @retrieve_student_classes_bp.route('/student/classes', methods=['GET'])
@@ -12,7 +12,7 @@ def retrieve_student_classes():
         student_id = request.args.get('student_id')
         print("Received student ID:", student_id)  # Debugging print
         if not student_id:
-            return jsonify(ClassesDatabase.generate_response(
+            return jsonify(StudentDatabase.generate_response(
                 success=False,
                 error='Missing student_id parameter',
                 status_code=400
@@ -20,13 +20,13 @@ def retrieve_student_classes():
 
         result = db.retrieve_student_classes(student_id)
         if not result['success']:
-            return jsonify(ClassesDatabase.generate_response(
+            return jsonify(StudentDatabase.generate_response(
                 success=False,
                 error=result['error'],
                 status_code=result['status_code']
             )), result['status_code']
 
-        return jsonify(ClassesDatabase.generate_response(
+        return jsonify(StudentDatabase.generate_response(
             success=True,
             data=result['data'],
             status_code=200
@@ -34,7 +34,7 @@ def retrieve_student_classes():
 
     except Exception as e:
         print("Exception occurred:", str(e))
-        return jsonify(ClassesDatabase.generate_response(
+        return jsonify(StudentDatabase.generate_response(
             success=False,
             error=str(e),
             status_code=500
