+++ routes/class_routes/retrieve_teacher_classes_route.py
@@ -1,9 +1,9 @@
 from flask import Blueprint, request, jsonify
-from modules.database_modules.class_database import ClassesDatabase
+from modules.database_modules.teacher_database import TeacherDatabase
 
 
 retrieve_teacher_classes_bp = Blueprint('retrieve_teacher_classes', __name__)
-db = ClassesDatabase()
+db = TeacherDatabase()
 
 
 @retrieve_teacher_classes_bp.route('/teacher/classes', methods=['GET'])
@@ -12,7 +12,7 @@ def retrieve_teacher_classes():
         teacher_id = request.args.get('teacher_id')
         print("Received teacher ID:", teacher_id)  # Debugging print
         if not teacher_id:
-            return jsonify(ClassesDatabase.generate_response(
+            return jsonify(TeacherDatabase.generate_response(
                 success=False,
                 error='Missing teacher_id parameter',
                 status_code=400
@@ -20,13 +20,13 @@ def retrieve_teacher_classes():
 
         result = db.retrieve_teacher_classes(teacher_id)
         if not result['success']:
-            return jsonify(ClassesDatabase.generate_response(
+            return jsonify(TeacherDatabase.generate_response(
                 success=False,
                 error=result['error'],
                 status_code=result['status_code']
             )), result['status_code']
 
-        return jsonify(ClassesDatabase.generate_response(
+        return jsonify(TeacherDatabase.generate_response(
             success=True,
             data=result['data'],
             status_code=200
@@ -34,7 +34,7 @@ def retrieve_teacher_classes():
 
     except Exception as e:
         print("Exception occurred:", str(e))
-        return jsonify(ClassesDatabase.generate_response(
+        return jsonify(TeacherDatabase.generate_response(
             success=False,
             error=str(e),
             status_code=500
